wallcon: Put in static_cast on int to workaround compilation error on raijin.
The GNU c++ library on raijin is somewhat old (4.4.7) and
does not provide the full set of overloaded functions
for the library function to_string(). As such, to_string()
applied to ints creates an ambiguity problem. The workaround
is to cast the int to long long.

+++ app/wallcon/source/useful_funcs.cxx
@@ -33,7 +33,7 @@ std::vector<double> read_time_varying_bc(std::string boundary, std::string dir, 
     std::string line;
 
     //Set the path
-    filename = dir + "/" + boundary + "/" +std::to_string(iteration) ;
+    filename = dir + "/" + boundary + "/" +std::to_string(static_cast<long long>(iteration)) ;
 
     //Read file with terms
 	std::ifstream user_file( filename.c_str() );
@@ -57,7 +57,7 @@ std::vector<double> read_time_varying_bc2(std::string bc_type, std::string bound
     std::string line;
 
     //Set the path
-    filename = dir + "/" + boundary + "/" +std::to_string(iteration) + "/" + bc_type ;
+    filename = dir + "/" + boundary + "/" +std::to_string(static_cast<long long>(iteration)) + "/" + bc_type ;
 
     //Read file with terms
 	std::ifstream user_file( filename.c_str() );
